+++ test/agent_configuration_test.dart
@@ -21,6 +21,7 @@ void main() {
     expect(config.screenshotsEnabled, config2.screenshotsEnabled);
     expect(config.crashReportCallback, config2.crashReportCallback);
     expect(config.crashReportingEnabled, config2.crashReportingEnabled);
+    expect(config.applicationName, config2.applicationName);
 
     AgentConfiguration config3 =
         config.copyWith(loggingLevel: LoggingLevel.info);